faceguard-v2/services/core-data-service/requirements.txt
# FACEGUARD V2 CORE DATA SERVICE DEPENDENCIES
# Rule 2: Zero Placeholder Code - Real production dependencies

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (LOCAL PostgreSQL)
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0

# Data Validation & Settings
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Vector Search (Replace Qdrant)
faiss-cpu==1.7.4

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional Dependencies
python-multipart==0.0.6
python-jose[cryptography]==3.3.0